+++ pitindex/__init__.py
@@ -31,4 +31,4 @@ __all__ = [
     "update",
 ]
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"